pyproject.toml
[tool.poetry]
name = "real-estate"
version = "0.1.0"
description = ""
authors = ["Andriy Orehov <andriyorehov@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"

# Core
pytz = "2020.1"
Django = "3.0.6"
django-configurations = "2.2"
django-extensions = "^2.2.9"
gunicorn = "20.0.4"

# For the persistence stores
psycopg2-binary = "2.8.6"
dj-database-url = "0.5.0"

# Model Tools
django-model-utils = "4.0.0"
django_unique_upload = "0.2.1"
django-cacheops = "^5.0"

# Rest apis
djangorestframework = "3.11.0"
Markdown = "3.2.1"
django-filter = "2.2.0"
drf-yasg = "^1.17.1"
djangorestframework-camel-case = "^1.1.2"
django-cors-headers = "^3.2.1"
drf-extensions = "^0.6.0"

# Developer Tools
ipdb = "0.13.2"
ipython = "7.14.0"
mock = "4.0.2"
flake8 = "3.7.9"

# Testing
mkdocs = "1.1"
factory-boy = "2.12.0"
django-nose = "1.4.6"
coverage = "5.1"


# Static and Media Storage
boto3 = "1.13.5"
dropbox = "^10.1.2"
django-environ = "^0.4.5"
newrelic = "^5.14.1"
django-storages = {git = "https://github.com/andriyor/django-storages.git", rev = "fix/dropbox-persist-url"}
drf-flex-fields = "^0.8.6"
httpx = "^0.16.1"
parsel = "^1.6.0"
djangorestframework-gis = "^0.16"
geojson = "^2.5.0"
django-modeltranslation = "^0.16.1"
django-debug-toolbar = "^3.1.1"
django-parler = "^2.2"
django-localized-fields = "^6.1"
django-postgres-extra = "^2.0.2"
django-parler-rest = "^2.1"
django-rest-multiple-models = "^2.1.3"
Pillow = "^8.4.0"
stone = "^3.2.1"
drf-extra-fields = "^3.2.1"

[tool.poetry.dev-dependencies]
pylint = "^2.6.0"

[build-system]
requires = ["poetry-core>=1.0.0a5"]
build-backend = "poetry.core.masonry.api"
